stress-io-uring: fix cast warning on 32 bit arm systems
Fix build warning:
stress-io-uring.c:273:19: warning: cast from pointer to integer
of different size [-Wpointer-to-int-cast]
  273 |  sqe->user_data = (unsigned long long)io_uring_file;

+++ stress-io-uring.c
@@ -267,10 +267,10 @@ static int stress_io_uring_iovec_submit(
 	sqe->fd = io_uring_file->fd;
 	sqe->flags = 0;
 	sqe->opcode = opcode;
-	sqe->addr = (unsigned long)io_uring_file->iovecs;
+	sqe->addr = (ptrdiff_t)io_uring_file->iovecs;
 	sqe->len = io_uring_file->blocks;
 	sqe->off = 0;
-	sqe->user_data = (unsigned long long)io_uring_file;
+	sqe->user_data = (ptrdiff_t)io_uring_file;
 	sring->array[index] = index;
 	tail = next_tail;
